Fixes (#759)
* uncomment spec test

* use a single db in tests to avoid memory overflow

* update state of highest instance when reading from db

* fix nil deref

* fix q clean

+++ protocol/v2/ssv/validator/msgqueue_consumer.go
@@ -87,7 +87,8 @@ func (v *Validator) ConsumeQueue(msgID spectypes.MessageID, handler MessageHandl
 
 		// clean all old messages. (when stuck on change round stage, msgs not deleted)
 		cleaned := v.Q.Clean(func(index msgqueue.Index) bool {
-			return index.H <= (lastHeight - 2) // remove all msg's that are 2 heights old. not post consensus & decided
+			// remove all msg's that are 2 heights old, besides height 0
+			return int64(index.H) <= int64(lastHeight-2) // remove all msg's that are 2 heights old. not post consensus & decided
 		})
 		if cleaned > 0 {
 			logger.Debug("indexes cleaned from queue", zap.Int64("count", cleaned))
